+++ audio-engine-instrument-bowed-string/src/damping.rs
@@ -1,22 +1,22 @@
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 
 use crate::string::String;
 
 pub trait DampingCoeffcient {
-    fn damping_coeffcient(&self, eigen_frequency: f32) -> f32;
-    fn damping_coeffcients(&self, eigen_frequencies: &[f32]) -> Vec<f32> {
+    fn damping_coeffcient(&self, eigen_frequency: f64) -> f64;
+    fn damping_coeffcients(&self, eigen_frequencies: &[f64]) -> Vec<f64> {
         eigen_frequencies
             .iter()
             .map(|eigen_frequency| self.damping_coeffcient(*eigen_frequency))
-            .collect::<Vec<f32>>()
+            .collect::<Vec<f64>>()
     }
 }
 
 impl DampingCoeffcient for String {
     /// Desvages damping coeffcients
-    fn damping_coeffcient(&self, eigen_frequency: f32) -> f32 {
-        const RHO_AIR: f32 = 1.225;
-        const MU_AIR: f32 = 1.619e-5;
+    fn damping_coeffcient(&self, eigen_frequency: f64) -> f64 {
+        const RHO_AIR: f64 = 1.225;
+        const MU_AIR: f64 = 1.619e-5;
 
         let d0 = -2.0 * RHO_AIR * MU_AIR / (self.density * self.radius * self.radius);
         let d1 = -2.0 * RHO_AIR * (2.0 * MU_AIR).sqrt() / (self.density * self.radius);
