+++ audio-engine-instrument-bowed-string/src/string_and_hand.rs
@@ -9,11 +9,11 @@ pub struct StringAndHand {
 impl StringAndHand {
     /// Get the actual length of the string that isn't blocked by the hand position.
     /// Length of the string with the hand blockage applied.
-    pub fn length(&self) -> f32 {
+    pub fn length(&self) -> f64 {
         self.string.length * self.hand.fretting_position
     }
 
-    pub fn upper_k(&self) -> f32 {
+    pub fn upper_k(&self) -> f64 {
         self.string.inertia() * self.string.young_mod
             / (self.string.lin_density() * self.length().powi(4)).sqrt()
     }
